CelestialObject

+++ CelestialObject.java
@@ -1,6 +1,6 @@
 public class CelestialObject {
-    double x=0.0;
-    double y=0.0;
-    double z=0.0;
+    Float x=0.0f;
+    Float y=0.0f;
+    Float z=0.0f;
     String name;
 }
